requirements.txt
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
mpmath==1.3.0
numpy==1.26.3
onnxruntime==1.16.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
pyreadline3==3.4.1
sympy==1.12
